requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=4.6.2=py312h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py312h5eee18b_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py312h06a4308_0
attrs=24.2.0=py312h06a4308_0
babel=2.11.0=py312h06a4308_0
beautifulsoup4=4.12.3=py312h06a4308_0
blas=1.0=mkl
bleach=6.2.0=py312h06a4308_0
blinker=1.6.2=py312h06a4308_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.11.26=h06a4308_0
certifi=2024.8.30=py312h06a4308_0
cffi=1.17.1=py312h1fdaa30_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312h06a4308_0
comm=0.2.1=py312h06a4308_0
contourpy=1.3.1=py312hdb19cb5_0
cpuonly=2.0=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.3=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
flask=3.0.3=py312h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py312h5eee18b_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.2=h5eee18b_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h11=0.14.0=py312h06a4308_0
httpcore=1.0.2=py312h06a4308_0
httpx=0.27.0=py312h06a4308_0
icu=73.1=h6a678d5_0
idna=3.7=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=py312h06a4308_0
ipython=8.27.0=py312h06a4308_0
ipywidgets=8.1.5=py312h06a4308_0
itsdangerous=2.2.0=py312h06a4308_0
jedi=0.19.1=py312h06a4308_0
jinja2=3.1.4=py312h06a4308_1
jpeg=9e=h5eee18b_3
json5=0.9.25=py312h06a4308_0
jsonschema=4.23.0=py312h06a4308_0
jsonschema-specifications=2023.7.1=py312h06a4308_0
jupyter=1.0.0=py312h06a4308_9
jupyter-lsp=2.2.0=py312h06a4308_0
jupyter_client=8.6.0=py312h06a4308_0
jupyter_console=6.6.3=py312h06a4308_1
jupyter_core=5.7.2=py312h06a4308_0
jupyter_events=0.10.0=py312h06a4308_0
jupyter_server=2.14.1=py312h06a4308_0
jupyter_server_terminals=0.4.4=py312h06a4308_1
jupyterlab=4.2.5=py312h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.27.3=py312h06a4308_0
jupyterlab_widgets=3.0.13=py312h06a4308_0
kiwisolver=1.4.4=py312h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.9.1=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hecde1de_4
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
libpq=17.0=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.5=hfdd30dd_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.9.2=py312h06a4308_1
matplotlib-base=3.9.2=py312hbfdbfaf_1
matplotlib-inline=0.1.6=py312h06a4308_0
mistune=2.0.4=py312h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.11=py312h5eee18b_0
mkl_random=1.2.8=py312h526ad5a_0
mpmath=1.3.0=py312h06a4308_0
mysql=8.4.0=h29a9f33_1
nbclient=0.8.0=py312h06a4308_0
nbconvert=7.16.4=py312h06a4308_0
nbformat=5.10.4=py312h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=py312h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py312h06a4308_0
notebook=7.2.2=py312h06a4308_1
notebook-shim=0.2.3=py312h06a4308_0
numpy=2.1.3=py312hc5e2394_0
numpy-base=2.1.3=py312h0da6c21_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openldap=2.6.4=h42fbc30_0
openssl=3.0.15=h5eee18b_0
overrides=7.4.0=py312h06a4308_0
packaging=24.1=py312h06a4308_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.42=hebb0a14_1
pexpect=4.8.0=pyhd3eb1b0_3
pillow=11.0.0=py312hfdbf927_0
pip=24.2=py312h06a4308_0
platformdirs=3.10.0=py312h06a4308_0
ply=3.11=py312h06a4308_1
prometheus_client=0.21.0=py312h06a4308_0
prompt-toolkit=3.0.43=py312h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py312h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py312h06a4308_1
pyparsing=3.2.0=py312h06a4308_0
pyqt=5.15.10=py312h6a678d5_0
pyqt5-sip=12.13.0=py312h5eee18b_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.7=h5148396_0
python-dateutil=2.9.0post0=py312h06a4308_2
python-fastjsonschema=2.20.0=py312h06a4308_0
python-json-logger=2.0.7=py312h06a4308_0
pytorch=2.5.1=py3.12_cpu_0
pytorch-mutex=1.0=cpu
pytz=2024.1=py312h06a4308_0
pyyaml=6.0.2=py312h5eee18b_0
pyzmq=25.1.2=py312h6a678d5_0
qt-main=5.15.2=hb6262e9_11
qtconsole=5.6.0=py312h06a4308_0
qtpy=2.4.1=py312h06a4308_0
readline=8.2=h5eee18b_0
referencing=0.30.2=py312h06a4308_0
requests=2.32.3=py312h06a4308_1
rfc3339-validator=0.1.4=py312h06a4308_0
rfc3986-validator=0.1.1=py312h06a4308_0
rpds-py=0.10.6=py312h4aa5aa6_1
send2trash=1.8.2=py312h06a4308_0
setuptools=72.1.0=py312h06a4308_0
sip=6.7.12=py312h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py312h06a4308_0
soupsieve=2.5=py312h06a4308_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.13.2=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
terminado=0.17.1=py312h06a4308_0
tinycss2=1.2.1=py312h06a4308_0
tk=8.6.14=h39e8969_0
torchaudio=2.5.1=py312_cpu
torchsummary=1.5.1=pypi_0
torchvision=0.20.1=py312_cpu
tornado=6.4.1=py312h5eee18b_0
tqdm=4.66.5=py312he106c6f_0
traitlets=5.14.3=py312h06a4308_0
typing-extensions=4.11.0=py312h06a4308_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024b=h04d1e81_0
unicodedata2=15.1.0=py312h5eee18b_0
urllib3=2.2.3=py312h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py312h06a4308_2
websocket-client=1.8.0=py312h06a4308_0
werkzeug=3.0.6=py312h06a4308_0
wheel=0.44.0=py312h06a4308_0
widgetsnbextension=4.0.13=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
